Removed file scripts loader test regions
